Move jsonitere outside of the internal package so it can be usable in grafana/* repos
